Replace aleph ws client with gniazdo (reasons below). Improve conns.
Aleph's abstractions around on-close and let-flow were not clear to me
and some subtle errors crept in. Gniazdo is a known quantity. I
originally switched to Aleph thinking that I would also switch the
server. That did not end up happening, so there is no real reason to
keep Aleph. I am sure it could be made to work; I just don't want to
spend the time on it when there are perfectly good alternatives.